backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
pyotp==2.9.0
openai==1.12.0
Werkzeug==3.0.1
cryptography==42.0.5
python-dotenv==1.0.1
psycopg[binary]
alembic
SQLAlchemy
clickhouse-sqlalchemy
pymysql
psycopg2-binary
